LayoutTests/platform/wpe/webgl/max-active-contexts-gc-expected.txt

CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: WebGL: INVALID_OPERATION: loseContext: context already lost
